Launch rocprofiler exporter with omniwatch
New configuration options to enable/disable rocprofiler, as well as
to configure the list of counters to sample.
