+++ omnistat/config/omnistat.default
@@ -6,6 +6,7 @@ usermode_port = 8001
 enable_rocm_smi = True
 enable_amd_smi = True
 enable_slurm = True
+enable_rocprofiler = False
 
 # Path to local ROCM install to access SMI library
 rocm_path = /opt/rocm-6.0.2
@@ -24,6 +25,10 @@ host_skip = "login.*"
 enable_annotations = False
 job_detection_mode = file-based
 job_detection_file = /tmp/omni_slurmjobinfo
+
+[omniwatch.collectors.rocprofiler]
+# comma-separated list of metric names
+metrics = GRBM_COUNT,SQ_WAVES
 
 [omnistat.report]
