+++ omnistat/monitor.py
@@ -58,6 +58,9 @@ class Monitor:
         self.runtimeConfig["collector_enable_amd_smi_process"] = config["omnistat.collectors"].getboolean(
             "enable_amd_smi_process", False
         )
+        self.runtimeConfig["collector_enable_rocprofiler"] = config["omnistat.collectors"].getboolean(
+            "enable_rocprofiler", False
+        )
         self.runtimeConfig["collector_port"] = config["omnistat.collectors"].get("port", 8000)
         self.runtimeConfig["collector_usermode"] = config["omnistat.collectors"].getboolean("usermode", False)
         self.runtimeConfig["collector_rocm_path"] = config["omnistat.collectors"].get("rocm_path", "/opt/rocm")
@@ -79,6 +82,9 @@ class Monitor:
             )
             if config.has_option("omnistat.collectors.slurm", "host_skip"):
                 self.runtimeConfig["slurm_collector_host_skip"] = config["omnistat.collectors.slurm"]["host_skip"]
+
+        if config.has_option("omniwatch.collectors.rocprofiler", "metrics"):
+            self.runtimeConfig["rocprofiler_metrics"] = config["omniwatch.collectors.rocprofiler"]["metrics"].split(",")
 
         # defined global prometheus metrics
         self.__globalMetrics = {}
@@ -124,6 +130,10 @@ class Monitor:
                     jobDetection=self.jobDetection,
                 )
             )
+        if self.runtimeConfig["collector_enable_rocprofiler"]:
+            from omnistat.collector_rocprofiler import rocprofiler
+
+            self.__collectors.append(rocprofiler(self.runtimeConfig['rocprofiler_metrics']))
 
         # Initialize all metrics
         for collector in self.__collectors:
